should work now

+++ pkg/plex/plex.go
@@ -17,8 +17,8 @@ type Server struct {
 	Name       string        `toml:"server"`
 	ReturnJSON bool          `toml:"return_json"`
 	Cooldown   cnfg.Duration `toml:"cooldown"`
-	SeriesPC   float64       `toml:"series_percent_complete"`
-	MoviesPC   float64       `toml:"movies_percent_complete"`
+	SeriesPC   uint          `toml:"series_percent_complete"`
+	MoviesPC   uint          `toml:"movies_percent_complete"`
 	client     *http.Client
 }
